requirements.txt
# Pinned versions (audited 2026-04-16). Update deliberately, not automatically.

# Discord listener
discord.py==2.7.1

# Telegram broadcaster + bot commands
python-telegram-bot==22.7

# Whop OAuth callback server (HTTP) and outbound HTTP client
aiohttp==3.13.5

# Verified-users persistence
aiosqlite==0.22.1

# Refresh-token encryption at rest
cryptography==46.0.7

# Config + secrets
PyYAML==6.0.3
python-dotenv==1.2.2

# Structured logging
structlog==25.5.0

# Testing
pytest==9.0.3
pytest-asyncio==1.3.0
